correct pointer arithmetic in triangle.cpp requires "unsigned int long long"
work-items: #57987, #57988
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Trunk;C58739

+++ 3rdparty/NNlib/src/triangle.cpp
@@ -659,7 +659,7 @@ REAL o3derrboundA, o3derrboundB, o3derrboundC;
 
 /* Random number seed is not constant, but I've made it global anyway.       */
 
-unsigned long randomseed;                     /* Current random number seed. */
+unsigned long long randomseed;                     /* Current random number seed. */
 
 
 /* Mesh data structure.  Triangle operates on only one mesh, but the mesh    */
@@ -835,7 +835,7 @@ struct behavior {
 /*  extracting an orientation (in the range 0 to 2) and a pointer to the     */
 /*  beginning of a triangle.  The encode() routine compresses a pointer to a */
 /*  triangle and an orientation into a single pointer.  My assumptions that  */
-/*  triangles are four-byte-aligned and that the `unsigned long' type is     */
+/*  triangles are four-byte-aligned and that the `unsigned long long' type is     */
 /*  long enough to hold a pointer are two of the few kludges in this program.*/
 /*                                                                           */
 /*  Subsegments are manipulated similarly.  A pointer to a subsegment        */
@@ -946,16 +946,16 @@ int minus1mod3[3] = {2, 0, 1};
 /*   extracted from the two least significant bits of the pointer.           */
 
 #define decode(ptr, otri)                                                     \
-  (otri).orient = (int) ((unsigned long) (ptr) & (unsigned long) 3l);         \
+  (otri).orient = (int) ((unsigned long long) (ptr) & (unsigned long long) 3l);         \
   (otri).tri = (triangle *)                                                   \
-                  ((unsigned long) (ptr) ^ (unsigned long) (otri).orient)
+                  ((unsigned long long) (ptr) ^ (unsigned long long) (otri).orient)
 
 /* encode() compresses an oriented triangle into a single pointer.  It       */
 /*   relies on the assumption that all triangles are aligned to four-byte    */
 /*   boundaries, so the two least significant bits of (otri).tri are zero.   */
 
 #define encode(otri)                                                          \
-  (triangle) ((unsigned long) (otri).tri | (unsigned long) (otri).orient)
+  (triangle) ((unsigned long long) (otri).tri | (unsigned long long) (otri).orient)
 
 /* The following handle manipulation primitives are all described by Guibas  */
 /*   and Stolfi.  However, Guibas and Stolfi use an edge-based data          */
@@ -1119,16 +1119,16 @@ int minus1mod3[3] = {2, 0, 1};
 
 #define infect(otri)                                                          \
   (otri).tri[6] = (triangle)                                                  \
-                    ((unsigned long) (otri).tri[6] | (unsigned long) 2l)
+                    ((unsigned long long) (otri).tri[6] | (unsigned long long) 2l)
 
 #define uninfect(otri)                                                        \
   (otri).tri[6] = (triangle)                                                  \
-                    ((unsigned long) (otri).tri[6] & ~ (unsigned long) 2l)
+                    ((unsigned long long) (otri).tri[6] & ~ (unsigned long long) 2l)
 
 /* Test a triangle for viral infection.                                      */
 
 #define infected(otri)                                                        \
-  (((unsigned long) (otri).tri[6] & (unsigned long) 2l) != 0l)
+  (((unsigned long long) (otri).tri[6] & (unsigned long long) 2l) != 0l)
 
 /* Check or set a triangle's attributes.                                     */
 
@@ -1166,16 +1166,16 @@ int minus1mod3[3] = {2, 0, 1};
 /*   are masked out to produce the real pointer.                             */
 
 #define sdecode(sptr, osub)                                                   \
-  (osub).ssorient = (int) ((unsigned long) (sptr) & (unsigned long) 1l);      \
+  (osub).ssorient = (int) ((unsigned long long) (sptr) & (unsigned long long) 1l);      \
   (osub).ss = (subseg *)                                                      \
-              ((unsigned long) (sptr) & ~ (unsigned long) 3l)
+              ((unsigned long long) (sptr) & ~ (unsigned long long) 3l)
 
 /* sencode() compresses an oriented subsegment into a single pointer.  It    */
 /*   relies on the assumption that all subsegments are aligned to two-byte   */
 /*   boundaries, so the least significant bit of (osub).ss is zero.          */
 
 #define sencode(osub)                                                         \
-  (subseg) ((unsigned long) (osub).ss | (unsigned long) (osub).ssorient)
+  (subseg) ((unsigned long long) (osub).ss | (unsigned long long) (osub).ssorient)
 
 /* ssym() toggles the orientation of a subsegment.                           */
 
@@ -3541,27 +3541,27 @@ struct otri *t;
   struct osub printsh;
   vertex printvertex;
 
-  fprintf(stderr, "triangle x%lx with orientation %d:\n", (unsigned long) t->tri,
+  fprintf(stderr, "triangle x%lx with orientation %d:\n", (unsigned long long) t->tri,
          t->orient);
   decode(t->tri[0], printtri);
   if (printtri.tri == m->dummytri) {
     fprintf(stderr, "    [0] = Outer space\n");
   } else {
-    fprintf(stderr, "    [0] = x%lx  %d\n", (unsigned long) printtri.tri,
+    fprintf(stderr, "    [0] = x%lx  %d\n", (unsigned long long) printtri.tri,
            printtri.orient);
   }
   decode(t->tri[1], printtri);
   if (printtri.tri == m->dummytri) {
     fprintf(stderr, "    [1] = Outer space\n");
   } else {
-    fprintf(stderr, "    [1] = x%lx  %d\n", (unsigned long) printtri.tri,
+    fprintf(stderr, "    [1] = x%lx  %d\n", (unsigned long long) printtri.tri,
            printtri.orient);
   }
   decode(t->tri[2], printtri);
   if (printtri.tri == m->dummytri) {
     fprintf(stderr, "    [2] = Outer space\n");
   } else {
-    fprintf(stderr, "    [2] = x%lx  %d\n", (unsigned long) printtri.tri,
+    fprintf(stderr, "    [2] = x%lx  %d\n", (unsigned long long) printtri.tri,
            printtri.orient);
   }
 
@@ -3570,37 +3570,37 @@ struct otri *t;
     fprintf(stderr, "    Origin[%d] = NULL\n", (t->orient + 1) % 3 + 3);
   else
     fprintf(stderr, "    Origin[%d] = x%lx  (%.12g, %.12g)\n",
-           (t->orient + 1) % 3 + 3, (unsigned long) printvertex,
+           (t->orient + 1) % 3 + 3, (unsigned long long) printvertex,
            printvertex[0], printvertex[1]);
   dest(*t, printvertex);
   if (printvertex == (vertex) NULL)
     fprintf(stderr, "    Dest  [%d] = NULL\n", (t->orient + 2) % 3 + 3);
   else
     fprintf(stderr, "    Dest  [%d] = x%lx  (%.12g, %.12g)\n",
-           (t->orient + 2) % 3 + 3, (unsigned long) printvertex,
+           (t->orient + 2) % 3 + 3, (unsigned long long) printvertex,
            printvertex[0], printvertex[1]);
   apex(*t, printvertex);
   if (printvertex == (vertex) NULL)
     fprintf(stderr, "    Apex  [%d] = NULL\n", t->orient + 3);
   else
     fprintf(stderr, "    Apex  [%d] = x%lx  (%.12g, %.12g)\n",
-           t->orient + 3, (unsigned long) printvertex,
+           t->orient + 3, (unsigned long long) printvertex,
            printvertex[0], printvertex[1]);
 
   if (b->usesegments) {
     sdecode(t->tri[6], printsh);
     if (printsh.ss != m->dummysub) {
-      fprintf(stderr, "    [6] = x%lx  %d\n", (unsigned long) printsh.ss,
+      fprintf(stderr, "    [6] = x%lx  %d\n", (unsigned long long) printsh.ss,
              printsh.ssorient);
     }
     sdecode(t->tri[7], printsh);
     if (printsh.ss != m->dummysub) {
-      fprintf(stderr, "    [7] = x%lx  %d\n", (unsigned long) printsh.ss,
+      fprintf(stderr, "    [7] = x%lx  %d\n", (unsigned long long) printsh.ss,
              printsh.ssorient);
     }
     sdecode(t->tri[8], printsh);
     if (printsh.ss != m->dummysub) {
-      fprintf(stderr, "    [8] = x%lx  %d\n", (unsigned long) printsh.ss,
+      fprintf(stderr, "    [8] = x%lx  %d\n", (unsigned long long) printsh.ss,
              printsh.ssorient);
     }
   }
@@ -3636,19 +3636,19 @@ struct osub *s;
   vertex printvertex;
 
   fprintf(stderr, "subsegment x%lx with orientation %d and mark %d:\n",
-         (unsigned long) s->ss, s->ssorient, mark(*s));
+         (unsigned long long) s->ss, s->ssorient, mark(*s));
   sdecode(s->ss[0], printsh);
   if (printsh.ss == m->dummysub) {
     fprintf(stderr, "    [0] = No subsegment\n");
   } else {
-    fprintf(stderr, "    [0] = x%lx  %d\n", (unsigned long) printsh.ss,
+    fprintf(stderr, "    [0] = x%lx  %d\n", (unsigned long long) printsh.ss,
            printsh.ssorient);
   }
   sdecode(s->ss[1], printsh);
   if (printsh.ss == m->dummysub) {
     fprintf(stderr, "    [1] = No subsegment\n");
   } else {
-    fprintf(stderr, "    [1] = x%lx  %d\n", (unsigned long) printsh.ss,
+    fprintf(stderr, "    [1] = x%lx  %d\n", (unsigned long long) printsh.ss,
            printsh.ssorient);
   }
 
@@ -3657,28 +3657,28 @@ struct osub *s;
     fprintf(stderr, "    Origin[%d] = NULL\n", 2 + s->ssorient);
   else
     fprintf(stderr, "    Origin[%d] = x%lx  (%.12g, %.12g)\n",
-           2 + s->ssorient, (unsigned long) printvertex,
+           2 + s->ssorient, (unsigned long long) printvertex,
            printvertex[0], printvertex[1]);
   sdest(*s, printvertex);
   if (printvertex == (vertex) NULL)
     fprintf(stderr, "    Dest  [%d] = NULL\n", 3 - s->ssorient);
   else
     fprintf(stderr, "    Dest  [%d] = x%lx  (%.12g, %.12g)\n",
-           3 - s->ssorient, (unsigned long) printvertex,
+           3 - s->ssorient, (unsigned long long) printvertex,
            printvertex[0], printvertex[1]);
 
   decode(s->ss[4], printtri);
   if (printtri.tri == m->dummytri) {
     fprintf(stderr, "    [4] = Outer space\n");
   } else {
-    fprintf(stderr, "    [4] = x%lx  %d\n", (unsigned long) printtri.tri,
+    fprintf(stderr, "    [4] = x%lx  %d\n", (unsigned long long) printtri.tri,
            printtri.orient);
   }
   decode(s->ss[5], printtri);
   if (printtri.tri == m->dummytri) {
     fprintf(stderr, "    [5] = Outer space\n");
   } else {
-    fprintf(stderr, "    [5] = x%lx  %d\n", (unsigned long) printtri.tri,
+    fprintf(stderr, "    [5] = x%lx  %d\n", (unsigned long long) printtri.tri,
            printtri.orient);
   }
 }
@@ -3709,7 +3709,7 @@ struct memorypool *pool;
 #endif /* not ANSI_DECLARATORS */
 
 {
-  unsigned long alignptr;
+  unsigned long long alignptr;
 
   pool->items = 0;
   pool->maxitems = 0;
@@ -3717,11 +3717,11 @@ struct memorypool *pool;
   /* Set the currently active block. */
   pool->nowblock = pool->firstblock;
   /* Find the first item in the pool.  Increment by the size of (VOID *). */
-  alignptr = (unsigned long) (pool->nowblock + 1);
+  alignptr = (unsigned long long) (pool->nowblock + 1);
   /* Align the item on an `alignbytes'-byte boundary. */
   pool->nextitem = (VOID *)
-    (alignptr + (unsigned long) pool->alignbytes -
-     (alignptr % (unsigned long) pool->alignbytes));
+    (alignptr + (unsigned long long) pool->alignbytes -
+     (alignptr % (unsigned long long) pool->alignbytes));
   /* There are lots of unallocated items left in this block. */
   pool->unallocateditems = pool->itemsperblock;
   /* The stack of deallocated items is empty. */
@@ -3832,7 +3832,7 @@ struct memorypool *pool;
 {
   VOID *newitem;
   VOID **newblock;
-  unsigned long alignptr;
+  unsigned long long alignptr;
 
   /* First check the linked list of dead items.  If the list is not   */
   /*   empty, allocate an item from the list rather than a fresh one. */
@@ -3855,11 +3855,11 @@ struct memorypool *pool;
       pool->nowblock = (VOID **) *(pool->nowblock);
       /* Find the first item in the block.    */
       /*   Increment by the size of (VOID *). */
-      alignptr = (unsigned long) (pool->nowblock + 1);
+      alignptr = (unsigned long long) (pool->nowblock + 1);
       /* Align the item on an `alignbytes'-byte boundary. */
       pool->nextitem = (VOID *)
-        (alignptr + (unsigned long) pool->alignbytes -
-         (alignptr % (unsigned long) pool->alignbytes));
+        (alignptr + (unsigned long long) pool->alignbytes -
+         (alignptr % (unsigned long long) pool->alignbytes));
       /* There are lots of unallocated items left in this block. */
       pool->unallocateditems = pool->itemsperblock;
     }
@@ -3917,16 +3917,16 @@ struct memorypool *pool;
 #endif /* not ANSI_DECLARATORS */
 
 {
-  unsigned long alignptr;
+  unsigned long long alignptr;
 
   /* Begin the traversal in the first block. */
   pool->pathblock = pool->firstblock;
   /* Find the first item in the block.  Increment by the size of (VOID *). */
-  alignptr = (unsigned long) (pool->pathblock + 1);
+  alignptr = (unsigned long long) (pool->pathblock + 1);
   /* Align with item on an `alignbytes'-byte boundary. */
   pool->pathitem = (VOID *)
-    (alignptr + (unsigned long) pool->alignbytes -
-     (alignptr % (unsigned long) pool->alignbytes));
+    (alignptr + (unsigned long long) pool->alignbytes -
+     (alignptr % (unsigned long long) pool->alignbytes));
   /* Set the number of items left in the current block. */
   pool->pathitemsleft = pool->itemsperblock;
 }
@@ -3954,7 +3954,7 @@ struct memorypool *pool;
 
 {
   VOID *newitem;
-  unsigned long alignptr;
+  unsigned long long alignptr;
 
   /* Stop upon exhausting the list of items. */
   if (pool->pathitem == pool->nextitem) {
@@ -3965,11 +3965,11 @@ struct memorypool *pool;
     /* Find the next block. */
     pool->pathblock = (VOID **) *(pool->pathblock);
     /* Find the first item in the block.  Increment by the size of (VOID *). */
-    alignptr = (unsigned long) (pool->pathblock + 1);
+    alignptr = (unsigned long long) (pool->pathblock + 1);
     /* Align with item on an `alignbytes'-byte boundary. */
     pool->pathitem = (VOID *)
-      (alignptr + (unsigned long) pool->alignbytes -
-       (alignptr % (unsigned long) pool->alignbytes));
+      (alignptr + (unsigned long long) pool->alignbytes -
+       (alignptr % (unsigned long long) pool->alignbytes));
     /* Set the number of items left in the current block. */
     pool->pathitemsleft = pool->itemsperblock;
   }
@@ -4024,16 +4024,16 @@ int subsegwords;
 #endif /* not ANSI_DECLARATORS */
 
 {
-  unsigned long alignptr;
+  unsigned long long alignptr;
 
   /* Set up `dummytri', the `triangle' that occupies "outer space." */
   m->dummytribase = (triangle *) trimalloc(trianglewords * sizeof(triangle) +
                                            m->triangles.alignbytes);
   /* Align `dummytri' on a `triangles.alignbytes'-byte boundary. */
-  alignptr = (unsigned long) m->dummytribase;
+  alignptr = (unsigned long long) m->dummytribase;
   m->dummytri = (triangle *)
-    (alignptr + (unsigned long) m->triangles.alignbytes -
-     (alignptr % (unsigned long) m->triangles.alignbytes));
+    (alignptr + (unsigned long long) m->triangles.alignbytes -
+     (alignptr % (unsigned long long) m->triangles.alignbytes));
   /* Initialize the three adjoining triangles to be "outer space."  These  */
   /*   will eventually be changed by various bonding operations, but their */
   /*   values don't really matter, as long as they can legally be          */
@@ -4053,10 +4053,10 @@ int subsegwords;
     m->dummysubbase = (subseg *) trimalloc(subsegwords * sizeof(subseg) +
                                            m->subsegs.alignbytes);
     /* Align `dummysub' on a `subsegs.alignbytes'-byte boundary. */
-    alignptr = (unsigned long) m->dummysubbase;
+    alignptr = (unsigned long long) m->dummysubbase;
     m->dummysub = (subseg *)
-      (alignptr + (unsigned long) m->subsegs.alignbytes -
-       (alignptr % (unsigned long) m->subsegs.alignbytes));
+      (alignptr + (unsigned long long) m->subsegs.alignbytes -
+       (alignptr % (unsigned long long) m->subsegs.alignbytes));
     /* Initialize the two adjoining subsegments to be the omnipresent      */
     /*   subsegment.  These will eventually be changed by various bonding  */
     /*   operations, but their values don't really matter, as long as they */
@@ -4406,7 +4406,7 @@ int number;
 {
   VOID **getblock;
   vertex foundvertex;
-  unsigned long alignptr;
+  unsigned long long alignptr;
   int current;
 
   getblock = m->vertices.firstblock;
@@ -4417,9 +4417,9 @@ int number;
     current += m->vertices.itemsperblock;
   }
   /* Now find the right vertex. */
-  alignptr = (unsigned long) (getblock + 1);
-  foundvertex = (vertex) (alignptr + (unsigned long) m->vertices.alignbytes -
-                          (alignptr % (unsigned long) m->vertices.alignbytes));
+  alignptr = (unsigned long long) (getblock + 1);
+  foundvertex = (vertex) (alignptr + (unsigned long long) m->vertices.alignbytes -
+                          (alignptr % (unsigned long long) m->vertices.alignbytes));
   while (current < number) {
     foundvertex += m->vertices.itemwords;
     current++;
@@ -6442,9 +6442,9 @@ struct mesh *m;
 /*****************************************************************************/
 
 #ifdef ANSI_DECLARATORS
-unsigned long randomnation(unsigned int choices)
+unsigned long long randomnation(unsigned int choices)
 #else /* not ANSI_DECLARATORS */
-unsigned long randomnation(choices)
+unsigned long long randomnation(choices)
 unsigned int choices;
 #endif /* not ANSI_DECLARATORS */
 
@@ -7637,7 +7637,7 @@ struct otri *searchtri;
   triangle *firsttri;
   struct otri sampletri;
   vertex torg, tdest;
-  unsigned long alignptr;
+  unsigned long long alignptr;
   REAL searchdist, dist;
   REAL ahead;
   long sampleblocks, samplesperblock, samplenum;
@@ -7694,9 +7694,9 @@ struct otri *searchtri;
   sampleblock = m->triangles.firstblock;
   sampletri.orient = 0;
   for (i = 0; i < sampleblocks; i++) {
-    alignptr = (unsigned long) (sampleblock + 1);
-    firsttri = (triangle *) (alignptr + (unsigned long) m->triangles.alignbytes
-                      - (alignptr % (unsigned long) m->triangles.alignbytes));
+    alignptr = (unsigned long long) (sampleblock + 1);
+    firsttri = (triangle *) (alignptr + (unsigned long long) m->triangles.alignbytes
+                      - (alignptr % (unsigned long long) m->triangles.alignbytes));
     for (j = 0; j < samplesperblock; j++) {
       if (i == triblocks - 1) {
         samplenum = randomnation((int)
